test-requirements.txt
appdirs==1.4.4
attrs==19.3.0
bandit==1.6.2
black==19.10b0
Click==7.1.1
Flask==1.1.2
gitdb==4.0.4
gitdb2==3.0.3
GitPython==3.1.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.2.0
mypy==0.770
mypy-extensions==0.4.3
packaging==20.3
pathspec==0.8.0
pbr==5.4.5
pluggy==0.13.1
py==1.8.1
pycryptodomex==3.9.7
pyparsing==2.4.7
pytest==5.4.1
python-gnupg==0.4.6
PyYAML==5.3.1
regex==2020.5.14
six==1.14.0
smmap==3.0.2
smmap2==3.0.1
stevedore==1.32.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
vulture==1.4
waitress==1.4.3
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
